Add timestamp and studio names to MIVS discussion report Fixes https://jira.magfest.net/browse/MAGDEV-630. Also splits the reports for MIVS out from the MIVS Admin section for easier navigation.
